avatar bug fix in navbar
