+++ frontend/src/stylesheets/frequency/frequency_game.scss
@@ -81,6 +81,7 @@
     justify-content: center;
     align-items: flex-end;
     margin: auto;
+    margin-top: 12px;
     text-align: center;
 }
