+++ frontend/src/stylesheets/navbar/dropdown.scss
@@ -19,7 +19,7 @@
      
     li:last-child {
         font-family: 'Really-Free';
-        color: $coral;
+        color: $text;
         font-size: 32px;
         margin: 12px 0px;
         padding: 12px 16px;
