+++ frontend/src/stylesheets/profile/avatar.scss
@@ -10,7 +10,7 @@
     bottom: 0;
     overflow: hidden;
     width: 100%;
-    background-color: rgba(0, 0, 0, .83);
+    background-color: rgba(0, 0, 0, .25);
     z-index: 20;
     transition: 0.2s;
 }
@@ -24,6 +24,7 @@
     top: 35%;
     left: 50%;
     transform: translate(-50%, 0);
+    // z-index: 1;
 }
 
 .profile-avatar-modal {
